Careful with intents in the web thread
Intents belong to the UI thread and they can be modified there.
So it's not prudent to use the activity's intent as a variable
in the code invoked from JavaScript. It's even more dangerous to
change the intent on the activity that's shared by all PhoneGap
plugins and may in fact be used in embedding code.